Alright alright alright. We're cooking now.
The current flow works like so:
- Render HTML on front end
- Do minor changes to make it fit for print.
- Send that HTML over to Flask
- Flask puts that message HTML into a base HTML file for letters
- We send that rendered HTML off to Lob. Whoo!

I hardcoded the heroku URL in lob_base.html because it has to be static
assets, it's ugly but it works for now since if I replace it with
`request.host` then we won't be able to upload letters when running
locally (the HTML we send to Lob would reference localhost:5000/ for the
CSS).

Cool? Cool.